Update test for legacy cached

+++ stress_cached_legacy_h.sh
@@ -43,6 +43,21 @@ usage() {
 	echo "  have to do anything."
 }
 
+init_binaries_and_folders() {
+	PITHOS_FOLDER=${ARCH_SCRIPTS}/pithos/pithos
+	ARCHIP_FOLDER=${ARCH_SCRIPTS}/pithos/archip
+	LOG_FOLDER=${ARCH_SCRIPTS}/log/stress_cached_legacy
+
+	LD_PRELOAD_PATH="LD_PRELOAD=${XSEG}/sys/user/libxseg.so"
+	XSEG_BIN="${LD_PRELOAD_PATH} ${XSEG}/peers/user/xseg"
+	BENCH_BIN="${LD_PRELOAD_PATH} ${XSEG}/peers/user/archip-bench"
+	CACHED_BIN="${LD_PRELOAD_PATH} ${XSEG}/peers/user/archip-cached"
+	PFILED_BIN="${LD_PRELOAD_PATH} ${XSEG}/peers/user/archip-pfiled"
+
+	# Create log folder
+	mkdir -p ${LOG_FOLDER}
+}
+
 parse_args() {
 	# ${1} is for threads
 	if [[ ${1} = 'single' ]]; then
@@ -157,7 +172,7 @@ print_test() {
 	eval "echo "${CACHED_COMMAND}""#"" \
 		| fmt -t -w 54 | sed -e 's/$/ \\/g' | sed -e 's/\# \\$//g'
 	echo ""
-	eval "echo "${MT_PFILED_COMMAND}""#"" \
+	eval "echo "${PFILED_COMMAND}""#"" \
 		| fmt -t -w 54 | sed -e 's/$/ \\/g' | sed -e 's/\# \\$//g'
 
 	grn_echo "-------------------------------------------------------"
@@ -165,7 +180,7 @@ print_test() {
 }
 
 init_log() {
-	LOG=/var/log/stress_cached_legacy/${1}
+	LOG=${LOG_FOLDER}/${1}
 
 	# Truncate previous logs
 	cat /dev/null > $LOG
@@ -191,14 +206,25 @@ restore_output() {
 }
 
 nuke_xseg() {
-	suppress_output
+	echo -n "Nuking xseg... "
 
-	# Delete mt-pfiled files
-	find /tmp/pithos1/ -name "*" -exec rm -rf {} \;
-	find /tmp/pithos2/ -name "*" -exec rm -rf {} \;
-	mkdir -p /tmp/pithos1/
-	mkdir -p /tmp/pithos2/
-	mkdir -p /var/log/stress_cached_legacy/
+	# Delete pfiled files
+	if [[ ! "$(basename $PITHOS_FOLDER)" = pithos ]] ||
+		[[ ! "$(basename $ARCHIP_FOLDER)" = archip ]]; then
+		red_echo "FAILED!"
+		echo ""
+		red_echo "There's something wrong with the pfiled folders"
+		red_echo "and you've just dodged a bullet..."
+		exit
+	fi
+
+	suppress_output
+	find ${PITHOS_FOLDER} -name "*" -exec rm -rf {} \;
+	find ${ARCHIP_FOLDER} -name "*" -exec rm -rf {} \;
+
+	# Re-build pfiled folders
+	mkdir -p ${PITHOS_FOLDER}
+	mkdir -p ${ARCHIP_FOLDER}
 
 	# Clear previous tries
 	killall -9 archip-bench
@@ -206,12 +232,17 @@ nuke_xseg() {
 	killall -9 archip-pfiled
 
 	# Re-build segment
-	xseg posix:apyrgio:16:1024:12 destroy create
+	eval $XSEG_BIN posix:cached:16:1024:12 destroy create
 	for P in $BENCH_PORTS; do
-		xseg posix:apyrgio: set-next ${P} 1
+		eval $XSEG_BIN posix:cached: set-next ${P} 1
 	done
-
 	restore_output
+
+	grn_echo "DONE!"
+}
+
+run_background() {
+	eval $(eval echo ${1})" &"
 }
 
 read_prompt () {
